feat(raiden): check cltv delay
This commit refines the logic to determine whether a Raiden payment
on the first leg of a swap (taker to maker) extends sufficient cltv
delay according to the maker's specified minimum cltv delay.

This also checks the `token` field of a Raiden resolve request to
ensure that it matches the expected token address for the deal.
It also improves calculation of the `cltvDeltaFactor` to use hardcoded
values for the average time per block for each chain, rather than using
the configured cltv delays for each currency. Changing the configured
values from the defaults could have resulted in a skewed factor.

The cltv delay for the final hop of a raiden payment is set to the
`expiration` value of the resolve request. Currently, given the limits
in Raiden for this value, the cltv delay for a swap where the first
leg of the swap uses Raiden will almost always be insufficient. The
only exception is if the configured cltv delay for the 2nd leg of
the swap is greatly reduced and if there are no hops.

Closes #1052.